refactoring only - moved enum & exception classes to their respective functional areas in the application
