+++ src/main/java/com/google/gcs/sdrs/ExecutionEventType.java
@@ -15,7 +15,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, and is not intended for production use.
  */
 
-package com.google.gcs.sdrs.enums;
+package com.google.gcs.sdrs;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gcs.sdrs.controller.validation.ValidationConstants;
