+++ src/main/java/com/google/gcs/sdrs/controller/PersistenceException.java
@@ -15,7 +15,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, and is not intended for production use.
  */
 
-package com.google.gcs.sdrs.controller.exception;
+package com.google.gcs.sdrs.controller;
 
 import javax.ws.rs.core.Response;
