+++ src/main/java/com/google/gcs/sdrs/controller/pojo/RetentionRuleUpdateResponse.java
@@ -17,7 +17,7 @@
 
 package com.google.gcs.sdrs.controller.pojo;
 
-import com.google.gcs.sdrs.enums.RetentionRuleType;
+import com.google.gcs.sdrs.RetentionRuleType;
 
 /** POJO Tracking JSON response fields/types when updating a retention rule */
 public class RetentionRuleUpdateResponse extends BaseHttpResponse {
