+++ src/main/java/com/google/gcs/sdrs/dao/converter/RetentionRuleTypeConverter.java
@@ -17,14 +17,15 @@
 
 package com.google.gcs.sdrs.dao.converter;
 
+import com.google.gcs.sdrs.RetentionRuleType;
 import com.google.gcs.sdrs.dao.util.DatabaseConstants;
-import com.google.gcs.sdrs.enums.RetentionRuleType;
+
+import static com.google.gcs.sdrs.RetentionRuleType.DATASET;
+import static com.google.gcs.sdrs.RetentionRuleType.GLOBAL;
+import static com.google.gcs.sdrs.RetentionRuleType.USER;
+
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
-
-import static com.google.gcs.sdrs.enums.RetentionRuleType.DATASET;
-import static com.google.gcs.sdrs.enums.RetentionRuleType.GLOBAL;
-import static com.google.gcs.sdrs.enums.RetentionRuleType.USER;
 
 /** Supports converting a RetentionRuleType between the database and application representations */
 @Converter
