+++ src/main/java/com/google/gcs/sdrs/dao/impl/RetentionRuleDaoImpl.java
@@ -17,9 +17,10 @@
 
 package com.google.gcs.sdrs.dao.impl;
 
+import com.google.gcs.sdrs.RetentionRuleType;
 import com.google.gcs.sdrs.dao.RetentionRuleDao;
 import com.google.gcs.sdrs.dao.model.RetentionRule;
-import com.google.gcs.sdrs.enums.RetentionRuleType;
+
 import java.util.LinkedList;
 import java.util.List;
 import javax.persistence.criteria.CriteriaBuilder;
