+++ src/main/java/com/google/gcs/sdrs/dao/model/RetentionRule.java
@@ -17,8 +17,9 @@
 
 package com.google.gcs.sdrs.dao.model;
 
+import com.google.gcs.sdrs.RetentionRuleType;
 import com.google.gcs.sdrs.dao.converter.RetentionRuleTypeConverter;
-import com.google.gcs.sdrs.enums.RetentionRuleType;
+
 import java.sql.Timestamp;
 import javax.persistence.Column;
 import javax.persistence.Convert;
