+++ src/main/java/com/google/gcs/sdrs/worker/impl/ExecuteRetentionWorker.java
@@ -17,13 +17,13 @@
 
 package com.google.gcs.sdrs.worker.impl;
 
+import com.google.gcs.sdrs.RetentionRuleType;
 import com.google.gcs.sdrs.controller.pojo.ExecutionEventRequest;
 import com.google.gcs.sdrs.dao.RetentionJobDao;
 import com.google.gcs.sdrs.dao.RetentionRuleDao;
 import com.google.gcs.sdrs.dao.SingletonDao;
 import com.google.gcs.sdrs.dao.model.RetentionJob;
 import com.google.gcs.sdrs.dao.model.RetentionRule;
-import com.google.gcs.sdrs.enums.RetentionRuleType;
 import com.google.gcs.sdrs.rule.RuleExecutor;
 import com.google.gcs.sdrs.rule.impl.StsRuleExecutor;
 import com.google.gcs.sdrs.util.RetentionUtil;
