+++ src/test/java/com/google/gcs/sdrs/controller/EventsControllerTest.java
@@ -17,13 +17,12 @@
 
 package com.google.gcs.sdrs.controller;
 
-import com.google.gcs.sdrs.controller.exception.HttpException;
-import com.google.gcs.sdrs.controller.exception.ValidationException;
+import com.google.gcs.sdrs.ExecutionEventType;
 import com.google.gcs.sdrs.controller.pojo.ErrorResponse;
 import com.google.gcs.sdrs.controller.pojo.EventResponse;
 import com.google.gcs.sdrs.controller.pojo.ExecutionEventRequest;
 import com.google.gcs.sdrs.controller.validation.ValidationResult;
-import com.google.gcs.sdrs.enums.ExecutionEventType;
+
 import javax.ws.rs.core.Response;
 
 import com.google.gcs.sdrs.service.impl.EventsServiceImpl;
