+++ src/test/java/com/google/gcs/sdrs/rule/impl/StsRuleValidatorTest.java
@@ -20,8 +20,8 @@ package com.google.gcs.sdrs.rule.impl;
 import com.google.api.client.googleapis.testing.auth.oauth2.MockGoogleCredential;
 import com.google.api.services.storagetransfer.v1.model.Operation;
 import com.google.api.services.storagetransfer.v1.model.Status;
+import com.google.gcs.sdrs.RetentionJobStatusType;
 import com.google.gcs.sdrs.dao.model.RetentionJobValidation;
-import com.google.gcs.sdrs.enums.RetentionJobStatusType;
 import com.google.gcs.sdrs.util.CredentialsUtil;
 import java.io.IOException;
 import java.util.HashMap;
