+++ src/test/java/com/google/gcs/sdrs/worker/impl/ValidationWorkerTest.java
@@ -7,10 +7,10 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.google.gcs.sdrs.RetentionJobStatusType;
 import com.google.gcs.sdrs.dao.RetentionJobValidationDao;
 import com.google.gcs.sdrs.dao.model.RetentionJob;
 import com.google.gcs.sdrs.dao.model.RetentionJobValidation;
-import com.google.gcs.sdrs.enums.RetentionJobStatusType;
 import com.google.gcs.sdrs.rule.impl.StsRuleValidator;
 import java.util.ArrayList;
 import java.util.List;
